src/ml/requirements.txt
fastapi
onnxruntime
numpy
pillow
uvicorn
python-multipart
